requirements.txt
streamlit==1.41.1
pydantic==2.10.5
python-dateutil==2.9.0.post0
